+++ grelmicro/task/__init__.py
@@ -1,7 +1,17 @@
 """Task."""
 
 from grelmicro.task._tasks import Tasks
-from grelmicro.task.errors import TaskError
+from grelmicro.task.errors import (
+    FunctionTypeError,
+    TaskAddOperationError,
+    TaskError,
+)
 from grelmicro.task.router import TaskRouter
 
-__all__ = ["TaskError", "TaskRouter", "Tasks"]
+__all__ = [
+    "FunctionTypeError",
+    "TaskAddOperationError",
+    "TaskError",
+    "TaskRouter",
+    "Tasks",
+]